[prometheus] Disk autoresizing removed
