+++ modules/300-prometheus/docs/FAQ_RU.md
@@ -456,3 +456,10 @@ spec:
   endpoints:
     - port: web
 ```
+
+## Как увеличить размер диска
+
+1. Для увеличения размера отредактируйте PersistentVolumeClaim, указав новый размер в поле `spec.resources.requests.storage`.
+    * Увеличение размера возможно если в StorageClass поле `allowVolumeExpansion` установлено в `true`.
+2. Когда размер диска будет изменен, в статусе PersistentVolumeClaim появится сообщение `Waiting for user to (re-)start a pod to finish file system resize of volume on node.`.
+3. Перезапустите Pod для завершения изменения размера файловой системы.
