+++ modules/300-prometheus/openapi/config-values.yaml
@@ -281,8 +281,12 @@ properties:
   mainMaxDiskSizeGigabytes:
     type: integer
     default: 300
-    description: The maximum size (in GiB) that the main Prometheus' volume can automatically resize to.
+    description: |
+      Deprecated and will be removed. Doesn't affect anything.
+      The maximum size (in GiB) that the main Prometheus' volume can automatically resize to.
   longtermMaxDiskSizeGigabytes:
     type: integer
     default: 300
-    description: the maximum size (in GiB) to which the Longterm Prometheus' disk can be automatically resized.
+    description: |
+      Deprecated and will be removed. Doesn't affect anything.
+      The maximum size (in GiB) to which the Longterm Prometheus' disk can be automatically resized.
