+++ modules/300-prometheus/openapi/doc-ru-config-values.yaml
@@ -154,6 +154,10 @@ properties:
 
       Если ничего не указано или указано `false` — будет [использоваться автоматика](https://deckhouse.io/ru/documentation/v1/#выделение-узлов-под-определенный-вид-нагрузки).
   mainMaxDiskSizeGigabytes:
-    description: Максимальный размер в гигабайтах, до которого автоматически может ресайзиться диск Prometheus.
+    description: |
+      Устарел и будет удален. Ни на что не влияет.
+      Максимальный размер в гигабайтах, до которого автоматически может ресайзиться диск Prometheus.
   longtermMaxDiskSizeGigabytes:
-    description: Максимальный размер в гигабайтах, до которого автоматически может ресайзиться диск Longterm Prometheus.
+    description: |
+      Устарел и будет удален. Ни на что не влияет.
+      Максимальный размер в гигабайтах, до которого автоматически может ресайзиться диск Longterm Prometheus.
